# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/customers": {
      "post": {
        "description": "Crea un registro asociado a la tienda del usuario autenticado. Devuelve el registro persistido con storeId y updatedAt, sin agregados calculados. Roles permitidos: OWNER, MANAGER, CASHIER.",
        "operationId": "CustomersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Operación completada (POST devuelve 201).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerRecord"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, propiedades no permitidas o regla de negocio incumplida. En usuarios, también se usa cuando el usuario no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente, inválido o vencido; usuario inexistente o inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto de unicidad: el valor ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno o de base de datos. Algunos errores Prisma no interceptados se devuelven con este estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear cliente",
        "tags": [
          "customers"
        ],
        "x-roles": [
          "OWNER",
          "MANAGER",
          "CASHIER"
        ]
      },
      "get": {
        "description": "Listado paginado por tienda. Busca por nombre, apellido, teléfono o correo. Orden predeterminado createdAt desc. Incluye numberOfOrders, amountSpent (suma de totales de todos sus pedidos) y canDelete. Roles permitidos: OWNER, MANAGER, CASHIER.",
        "operationId": "CustomersController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Página solicitada; enviar un entero positivo",
            "schema": {
              "minimum": 0,
              "exclusiveMinimum": true,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Elementos por página; enviar un entero positivo",
            "schema": {
              "minimum": 0,
              "exclusiveMinimum": true,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Texto de búsqueda, sin distinguir mayúsculas",
            "schema": {
              "example": "cafe",
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Campo del modelo Prisma para ordenar; no usar nombres calculados de la respuesta",
            "schema": {
              "default": "createdAt",
              "example": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "direction",
            "required": false,
            "in": "query",
            "description": "Dirección del orden",
            "schema": {
              "default": "desc",
              "example": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación completada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerPage"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, propiedades no permitidas o regla de negocio incumplida. En usuarios, también se usa cuando el usuario no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente, inválido o vencido; usuario inexistente o inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno o de base de datos. Algunos errores Prisma no interceptados se devuelven con este estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar clientes",
        "tags": [
          "customers"
        ],
        "x-roles": [
          "OWNER",
          "MANAGER",
          "CASHIER"
        ]
      }
    },
    "/api/customers/{id}": {
      "get": {
        "description": "Obtiene el registro identificado por UUID dentro de la tienda del usuario. Incluye numberOfOrders, amountSpent (suma de totales de todos sus pedidos) y canDelete. Roles permitidos: OWNER, MANAGER, CASHIER.",
        "operationId": "CustomersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso en la tienda del usuario.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación completada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, propiedades no permitidas o regla de negocio incumplida. En usuarios, también se usa cuando el usuario no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente, inválido o vencido; usuario inexistente o inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado en la tienda del usuario o registro relacionado inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno o de base de datos. Algunos errores Prisma no interceptados se devuelven con este estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Consultar cliente",
        "tags": [
          "customers"
        ],
        "x-roles": [
          "OWNER",
          "MANAGER",
          "CASHIER"
        ]
      },
      "patch": {
        "description": "Actualización parcial: las propiedades omitidas conservan su valor. Devuelve el registro persistido con storeId y updatedAt, sin agregados calculados. Roles permitidos: OWNER, MANAGER, CASHIER.",
        "operationId": "CustomersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso en la tienda del usuario.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operación completada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerRecord"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, propiedades no permitidas o regla de negocio incumplida. En usuarios, también se usa cuando el usuario no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente, inválido o vencido; usuario inexistente o inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado en la tienda del usuario o registro relacionado inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto de unicidad: el valor ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno o de base de datos. Algunos errores Prisma no interceptados se devuelven con este estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar cliente",
        "tags": [
          "customers"
        ],
        "x-roles": [
          "OWNER",
          "MANAGER",
          "CASHIER"
        ]
      },
      "delete": {
        "description": "Elimina el registro y devuelve sus datos. Las relaciones existentes pueden impedir la eliminación. Devuelve el registro persistido con storeId y updatedAt, sin agregados calculados. Si tiene pedidos devuelve 403. Roles permitidos: OWNER, MANAGER.",
        "operationId": "CustomersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso en la tienda del usuario.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación completada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerRecord"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, propiedades no permitidas o regla de negocio incumplida. En usuarios, también se usa cuando el usuario no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente, inválido o vencido; usuario inexistente o inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Cliente con pedidos asociados: no se puede eliminar.\n\nRequiere uno de estos roles: OWNER, MANAGER.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado en la tienda del usuario o registro relacionado inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto de unicidad: el valor ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno o de base de datos. Algunos errores Prisma no interceptados se devuelven con este estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar cliente",
        "tags": [
          "customers"
        ],
        "x-roles": [
          "OWNER",
          "MANAGER"
        ]
      }
    },
...
  },
  "info": {
    "title": "La Hermandad — API POS",
    "description": "API del punto de venta: autenticación, usuarios, categorías, productos, clientes y pedidos con pagos. Las rutas protegidas obtienen la tienda desde el usuario autenticado. Use login o register y pegue el token en Authorize. Los JWT duran 24 horas. Los cuerpos rechazan propiedades desconocidas. Los importes de entrada son números y los Decimal de salida son cadenas. Todos los POST devuelven 201; PATCH y DELETE devuelven 200 con cuerpo. Los permisos se indican mediante x-roles en cada operación.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Pegue únicamente el token obtenido en auth/login o auth/register."
      }
    },
    "schemas": {
...
      "CreateCustomerDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "Nombre del usuario o cliente",
            "example": "Álvaro",
            "minLength": 3
          },
          "lastName": {
            "type": "string",
            "description": "Apellidos",
            "example": "Pérez",
            "minLength": 3
          },
          "email": {
            "type": "string",
            "description": "Correo electrónico",
            "example": "alvaro@example.com",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "description": "Teléfono en formato internacional",
            "example": "+59171234567"
          },
          "note": {
            "type": "string",
            "description": "Nota interna",
            "example": "Atender por la tarde"
          }
        },
        "required": [
          "firstName"
        ]
      },
      "UpdateCustomerDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "Nombre del usuario o cliente",
            "example": "Álvaro",
            "minLength": 3
          },
          "lastName": {
            "type": "string",
            "description": "Apellidos",
            "example": "Pérez",
            "minLength": 3
          },
          "email": {
            "type": "string",
            "description": "Correo electrónico",
            "example": "alvaro@example.com",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "description": "Teléfono en formato internacional",
            "example": "+59171234567"
          },
          "note": {
            "type": "string",
            "description": "Nota interna",
            "example": "Atender por la tarde"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "statusCode",
          "message"
        ],
        "properties": {
          "statusCode": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "description": "Mensaje del servicio o lista de errores de validación.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": [
              "property unknown should not exist"
            ]
          },
          "error": {
            "type": "string",
            "description": "Nombre HTTP del error, cuando Nest lo incluye",
            "example": "Bad Request"
          }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "required": [
          "totalItems",
          "currentPage",
          "pageSize",
          "totalPages",
          "hasNextPage",
          "hasPreviousPage"
        ],
        "properties": {
          "totalItems": {
            "type": "integer",
            "example": 1
          },
          "currentPage": {
            "type": "integer",
            "example": 1
          },
          "pageSize": {
            "type": "integer",
            "example": 10
          },
          "totalPages": {
            "type": "integer",
            "example": 1
          },
          "hasNextPage": {
            "type": "boolean",
            "example": true
          },
          "hasPreviousPage": {
            "type": "boolean",
            "example": false
          }
        }
      },
...
      "CustomerRecord": {
        "type": "object",
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "phone",
          "createdAt",
          "storeId",
          "note",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-42d3-a456-426614174000"
          },
          "firstName": {
            "type": "string",
            "description": "Nombre",
            "example": "Álvaro"
          },
          "lastName": {
            "type": "string",
            "description": "Apellidos",
            "example": "Pérez",
            "nullable": true
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "alvaro@example.com",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "description": "Teléfono internacional",
            "example": "+59171234567",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-27T12:00:00.000Z"
          },
          "storeId": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-42d3-a456-426614174000"
          },
          "note": {
            "type": "string",
            "description": "Nota",
            "example": "Ejemplo",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-27T12:00:00.000Z"
          }
        }
      },
      "Customer": {
        "type": "object",
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "phone",
          "createdAt",
          "displayName",
          "note",
          "amountSpent",
          "numberOfOrders",
          "canDelete"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-42d3-a456-426614174000"
          },
          "firstName": {
            "type": "string",
            "description": "Nombre",
            "example": "Álvaro"
          },
          "lastName": {
            "type": "string",
            "description": "Apellidos",
            "example": "Pérez",
            "nullable": true
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "alvaro@example.com",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "description": "Teléfono internacional",
            "example": "+59171234567",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-27T12:00:00.000Z"
          },
          "displayName": {
            "type": "string",
            "description": "Nombre completo",
            "example": "Álvaro Pérez"
          },
          "note": {
            "type": "string",
            "description": "Nota",
            "example": "Ejemplo",
            "nullable": true
          },
          "amountSpent": {
            "type": "string",
            "description": "Decimal de Prisma serializado como cadena JSON; no garantiza ceros finales.",
            "example": "25.5"
          },
          "numberOfOrders": {
            "type": "integer",
            "example": 1
          },
          "canDelete": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "CustomerPage": {
        "type": "object",
        "required": [
          "meta",
          "results"
        ],
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Customer"
            }
          }
        }
      },
...
    }
  }
}